# bins/pg_kronika-web/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "PgKronika machine API",
    "version": "0.0.0",
    "summary": "Pre-release locale-neutral API contract"
  },
...
  "paths": {
...
    "/v1/sections": {
      "get": {
        "operationId": "sections",
        "security": [{ "basicAuth": [] }, {}],
        "responses": {
          "200": {
            "description": "Stable registry metadata",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Sections" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ApplicationProblem" },
          "401": { "$ref": "#/components/responses/ApplicationProblem" },
          "413": { "$ref": "#/components/responses/ApplicationProblem" },
          "500": { "$ref": "#/components/responses/ApplicationProblem" }
        }
      }
    },
...
    "/v1/anomalies": {
      "get": {
        "operationId": "anomalies",
        "security": [{ "basicAuth": [] }, {}],
        "parameters": [
          { "$ref": "#/components/parameters/source" },
          { "$ref": "#/components/parameters/from" },
          { "$ref": "#/components/parameters/to" },
          { "$ref": "#/components/parameters/window" },
          { "$ref": "#/components/parameters/step" },
          { "$ref": "#/components/parameters/threshold" },
          { "$ref": "#/components/parameters/epsRel" },
          { "$ref": "#/components/parameters/limit" },
          { "$ref": "#/components/parameters/section" }
        ],
        "responses": {
          "200": {
            "description": "Locale-neutral anomaly result with typed reasons",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AnomalyResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ApplicationProblem" },
          "401": { "$ref": "#/components/responses/ApplicationProblem" },
          "404": { "$ref": "#/components/responses/ApplicationProblem" },
          "413": { "$ref": "#/components/responses/ApplicationProblem" },
          "500": { "$ref": "#/components/responses/ApplicationProblem" },
          "503": { "$ref": "#/components/responses/ApplicationProblem" }
        }
      }
    },
    "/v1/incidents": {
      "get": {
        "operationId": "incidents",
        "security": [{ "basicAuth": [] }, {}],
        "parameters": [
          { "$ref": "#/components/parameters/source" },
          { "$ref": "#/components/parameters/from" },
          { "$ref": "#/components/parameters/to" },
          { "$ref": "#/components/parameters/window" },
          { "$ref": "#/components/parameters/step" },
          { "$ref": "#/components/parameters/threshold" },
          { "$ref": "#/components/parameters/epsRel" },
          { "$ref": "#/components/parameters/epsilon" },
          { "$ref": "#/components/parameters/maxClusterSpan" },
          { "$ref": "#/components/parameters/section" }
        ],
        "responses": {
          "200": {
            "description": "Locale-neutral incidents, evidence, catalogs, and typed reasons",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/IncidentResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ApplicationProblem" },
          "401": { "$ref": "#/components/responses/ApplicationProblem" },
          "404": { "$ref": "#/components/responses/ApplicationProblem" },
          "413": { "$ref": "#/components/responses/ApplicationProblem" },
          "500": { "$ref": "#/components/responses/ApplicationProblem" },
          "503": { "$ref": "#/components/responses/ApplicationProblem" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "parameters": {
...
      "source": { "name": "source", "in": "query", "required": true, "schema": { "type": "integer", "minimum": 0 } },
      "from": { "name": "from", "in": "query", "required": true, "schema": { "type": "integer", "format": "int64" } },
      "to": { "name": "to", "in": "query", "required": true, "schema": { "type": "integer", "format": "int64" } },
      "window": { "name": "window", "in": "query", "required": false, "schema": { "type": "string" } },
      "step": { "name": "step", "in": "query", "required": false, "schema": { "type": "string" } },
      "threshold": { "name": "threshold", "in": "query", "required": false, "schema": { "type": "number", "minimum": 0 } },
      "epsRel": { "name": "eps_rel", "in": "query", "required": false, "schema": { "type": "number", "minimum": 0 } },
      "epsilon": { "name": "epsilon", "in": "query", "required": false, "schema": { "type": "string" } },
      "maxClusterSpan": { "name": "max_cluster_span", "in": "query", "required": false, "schema": { "type": "string" } },
      "section": { "name": "section", "in": "query", "required": false, "schema": { "type": "string" } },
...
      "limit": { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 0, "maximum": 10000 } },
...
    },
    "headers": {
      "ProblemCacheControl": {
        "description": "Always no-store",
        "schema": { "type": "string", "const": "no-store" }
      },
      "ProblemRequestId": {
        "description": "Opaque server-generated correlation id used by instance",
        "schema": { "type": "string", "pattern": "^[0-9a-f]{32}$" }
      }
    },
    "responses": {
      "ApplicationProblem": {
        "description": "Locale-neutral RFC 9457 application error",
        "headers": {
          "Cache-Control": { "$ref": "#/components/headers/ProblemCacheControl" },
          "X-Request-ID": { "$ref": "#/components/headers/ProblemRequestId" }
        },
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      }
    },
    "schemas": {
...
      "Sections": {
        "type": "object",
        "additionalProperties": false,
        "required": ["sections"],
        "properties": {
          "sections": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SectionCatalogEntry" }
          }
        }
      },
      "SectionCatalogEntry": {
        "type": "object",
        "additionalProperties": false,
        "required": ["name", "semantics", "sort_key", "columns"],
        "properties": {
          "name": { "type": "string" },
          "semantics": {
            "type": "string",
            "enum": ["snapshot_full", "conditional_full", "event_stream", "changed", "on_change"]
          },
          "sort_key": { "type": "array", "items": { "type": "string" } },
          "columns": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SectionColumn" }
          }
        }
      },
      "SectionColumn": {
        "type": "object",
        "additionalProperties": false,
        "required": ["name", "type", "class"],
        "properties": {
          "name": { "type": "string" },
          "type": {
            "type": "string",
            "enum": ["i8", "i16", "i32", "i64", "u8", "u16", "u32", "u64", "f32", "f64", "bool", "ts", "str", "list_i32"]
          },
          "class": { "type": "string", "enum": ["c", "g", "l", "t"] }
        }
      },
...
      "NeutralObject": {
        "type": "object",
        "description": "Machine data only; raw source and user strings remain literal"
      },
      "AnomalyResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["source_id", "from", "to", "window_us", "step_us", "threshold", "eps_rel", "limit", "episodes", "sections", "skipped"],
        "properties": {
          "source_id": { "type": "integer", "minimum": 0 },
          "from": { "type": "integer", "format": "int64" },
          "to": { "type": "integer", "format": "int64" },
          "window_us": { "type": "integer", "format": "int64", "minimum": 1 },
          "step_us": { "type": "integer", "format": "int64", "minimum": 1 },
          "threshold": { "type": "number", "minimum": 0 },
          "eps_rel": { "type": "number", "minimum": 0 },
          "limit": { "type": "integer", "minimum": 0 },
          "episodes": { "type": "array", "items": { "$ref": "#/components/schemas/NeutralObject" } },
          "sections": { "type": "object", "additionalProperties": { "$ref": "#/components/schemas/NeutralObject" } },
          "skipped": { "type": "array", "items": { "$ref": "#/components/schemas/SectionReason" } }
        }
      },
      "SectionReason": {
        "type": "object",
        "additionalProperties": false,
        "required": ["section", "reason"],
        "properties": {
          "section": { "type": "string" },
          "reason": { "$ref": "#/components/schemas/Reason" }
        }
      },
      "IncidentResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["source_id", "from", "to", "complete", "clustering_complete", "analysis_status", "incidents", "coverage_by_section", "data_age_seconds", "catalog", "log", "data_quality", "skipped"],
        "properties": {
          "source_id": { "type": "integer", "minimum": 0 },
          "from": { "type": "integer", "format": "int64" },
          "to": { "type": "integer", "format": "int64" },
          "complete": { "type": "boolean" },
          "clustering_complete": { "type": "boolean" },
          "analysis_status": { "type": "string", "enum": ["no_data", "partial", "insufficient_data", "calm", "incidents_detected", "missing_node_identity", "conflicting_node_identity"] },
          "incidents": { "type": "array", "items": { "$ref": "#/components/schemas/NeutralObject" } },
          "coverage_by_section": { "type": "object", "additionalProperties": { "$ref": "#/components/schemas/NeutralObject" } },
          "data_age_seconds": { "type": ["integer", "null"], "minimum": 0 },
          "catalog": { "$ref": "#/components/schemas/IncidentCatalog" },
          "log": { "$ref": "#/components/schemas/IncidentLog" },
          "data_quality": { "$ref": "#/components/schemas/NeutralObject" },
          "skipped": { "$ref": "#/components/schemas/IncidentSkipped" }
        }
      },
      "IncidentCatalog": {
        "type": "object",
        "additionalProperties": false,
        "required": ["schema_version", "status", "requirements_status", "diagnosis_available", "scope", "applied", "active_count", "catalog_count", "capabilities", "dormant"],
        "properties": {
          "schema_version": { "type": "integer", "const": 1 },
          "status": { "type": "string", "const": "partial" },
          "requirements_status": { "type": "string", "const": "incomplete" },
          "diagnosis_available": { "type": "boolean" },
          "scope": { "type": "string", "const": "diagnostic_lenses" },
          "applied": { "type": "array", "items": { "type": "string" } },
          "active_count": { "type": "integer", "minimum": 0 },
          "catalog_count": { "type": "integer", "minimum": 0 },
          "capabilities": { "type": "array", "items": { "$ref": "#/components/schemas/LensCapability" } },
          "dormant": { "type": "array", "items": { "$ref": "#/components/schemas/DormantLens" } }
        }
      },
...
      "IncidentLog": {
        "type": "object",
        "additionalProperties": false,
        "required": ["schema_version", "complete", "evaluated_lens_ids", "catalog", "findings", "coverage", "skipped"],
        "properties": {
          "schema_version": { "type": "integer", "const": 1 },
          "complete": { "type": "boolean" },
          "evaluated_lens_ids": { "type": "array", "items": { "type": "string" } },
          "catalog": { "type": "array", "items": { "$ref": "#/components/schemas/EventCatalogEntry" } },
          "findings": { "type": "array", "items": { "$ref": "#/components/schemas/NeutralObject" } },
          "coverage": { "type": "object", "additionalProperties": { "type": "string" } },
          "skipped": { "type": "array", "items": { "$ref": "#/components/schemas/EngineSkip" } }
        }
      },
...
      "IncidentSkipped": {
        "type": "object",
        "additionalProperties": false,
        "required": ["sections", "evaluations", "analysis", "span_splits"],
        "properties": {
          "sections": { "type": "array", "items": { "$ref": "#/components/schemas/SectionReason" } },
          "evaluations": { "type": "array", "items": { "$ref": "#/components/schemas/EngineSkip" } },
          "analysis": { "type": "array", "items": { "$ref": "#/components/schemas/AnalysisReason" } },
          "span_splits": { "type": "integer", "minimum": 0 }
        }
      },
...
      "EmptyParams": {
        "type": "object",
        "additionalProperties": false,
        "maxProperties": 0
      },
      "ParameterParams": {
        "type": "object",
        "additionalProperties": false,
        "required": ["parameter"],
        "properties": { "parameter": { "$ref": "#/components/schemas/KnownParameter" } }
      },
      "InvalidParameterParams": {
        "type": "object",
        "additionalProperties": false,
        "required": ["parameter", "expected"],
        "properties": {
          "parameter": { "$ref": "#/components/schemas/AnyParameterLocation" },
          "expected": {
            "type": "string",
            "enum": ["url_encoded_query", "uint64", "int64", "positive_duration", "non_negative_finite_number", "non_negative_integer", "section_list"]
          }
        }
      },
      "UnknownParameterParams": {
        "type": "object",
        "additionalProperties": false,
        "required": ["parameter"],
        "properties": {
          "parameter": { "type": "string", "pattern": "^[A-Za-z0-9_.-]{1,64}$" }
        }
      },
      "ConstraintParams": {
        "type": "object",
        "additionalProperties": false,
        "required": ["constraint"],
        "properties": {
          "constraint": {
            "type": "string",
            "enum": ["from_before_to", "window_within_interval", "epsilon_not_greater_than_max_cluster_span", "max_cluster_span_within_interval", "finite_scan"]
          }
        }
      },
      "SectionParams": {
        "type": "object",
        "additionalProperties": false,
        "required": ["section"],
        "properties": {
          "section": { "type": "string", "pattern": "^[A-Za-z0-9_.-]{1,64}$" }
        }
      },
...
      "CapacityParams": {
        "type": "object",
        "additionalProperties": false,
        "required": ["retry_after_seconds"],
        "properties": { "retry_after_seconds": { "type": "integer", "const": 1 } }
      },
      "Problem": {
        "$id": "https://pgkronika.dev/schemas/problem.json",
        "type": "object",
        "additionalProperties": false,
        "required": ["type", "status", "code", "params", "instance"],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "https://pgkronika.dev/problems/unauthorized",
              "https://pgkronika.dev/problems/route-not-found",
              "https://pgkronika.dev/problems/method-not-allowed",
              "https://pgkronika.dev/problems/missing-query-parameter",
              "https://pgkronika.dev/problems/invalid-query-parameter",
              "https://pgkronika.dev/problems/unknown-query-parameter",
              "https://pgkronika.dev/problems/duplicate-query-parameter",
              "https://pgkronika.dev/problems/invalid-query-constraint",
              "https://pgkronika.dev/problems/unknown-section",
              "https://pgkronika.dev/problems/invalid-cursor",
              "https://pgkronika.dev/problems/query-limit-exceeded",
              "https://pgkronika.dev/problems/analytic-capacity-unavailable",
              "https://pgkronika.dev/problems/store-read-failed",
              "https://pgkronika.dev/problems/internal-error"
            ]
          },
          "status": { "type": "integer", "enum": [400, 401, 404, 405, 413, 500, 503] },
          "code": {
            "type": "string",
            "enum": ["unauthorized", "route_not_found", "method_not_allowed", "missing_query_parameter", "invalid_query_parameter", "unknown_query_parameter", "duplicate_query_parameter", "invalid_query_constraint", "unknown_section", "invalid_cursor", "query_limit_exceeded", "analytic_capacity_unavailable", "store_read_failed", "internal_error"]
          },
          "params": { "type": "object" },
          "instance": { "type": "string", "pattern": "^urn:pgkronika:request:[0-9a-f]{32}$" }
        },
        "allOf": [
          { "if": { "properties": { "code": { "const": "unauthorized" } } }, "then": { "properties": { "type": { "const": "https://pgkronika.dev/problems/unauthorized" }, "status": { "const": 401 }, "params": { "$ref": "#/components/schemas/EmptyParams" } } } },
          { "if": { "properties": { "code": { "const": "route_not_found" } } }, "then": { "properties": { "type": { "const": "https://pgkronika.dev/problems/route-not-found" }, "status": { "const": 404 }, "params": { "$ref": "#/components/schemas/EmptyParams" } } } },
          { "if": { "properties": { "code": { "const": "method_not_allowed" } } }, "then": { "properties": { "type": { "const": "https://pgkronika.dev/problems/method-not-allowed" }, "status": { "const": 405 }, "params": { "$ref": "#/components/schemas/EmptyParams" } } } },
          { "if": { "properties": { "code": { "const": "missing_query_parameter" } } }, "then": { "properties": { "type": { "const": "https://pgkronika.dev/problems/missing-query-parameter" }, "status": { "const": 400 }, "params": { "$ref": "#/components/schemas/ParameterParams" } } } },
          { "if": { "properties": { "code": { "const": "invalid_query_parameter" } } }, "then": { "properties": { "type": { "const": "https://pgkronika.dev/problems/invalid-query-parameter" }, "status": { "const": 400 }, "params": { "$ref": "#/components/schemas/InvalidParameterParams" } } } },
          { "if": { "properties": { "code": { "const": "unknown_query_parameter" } } }, "then": { "properties": { "type": { "const": "https://pgkronika.dev/problems/unknown-query-parameter" }, "status": { "const": 400 }, "params": { "$ref": "#/components/schemas/UnknownParameterParams" } } } },
          { "if": { "properties": { "code": { "const": "duplicate_query_parameter" } } }, "then": { "properties": { "type": { "const": "https://pgkronika.dev/problems/duplicate-query-parameter" }, "status": { "const": 400 }, "params": { "$ref": "#/components/schemas/ParameterParams" } } } },
          { "if": { "properties": { "code": { "const": "invalid_query_constraint" } } }, "then": { "properties": { "type": { "const": "https://pgkronika.dev/problems/invalid-query-constraint" }, "status": { "const": 400 }, "params": { "$ref": "#/components/schemas/ConstraintParams" } } } },
          { "if": { "properties": { "code": { "const": "unknown_section" } } }, "then": { "properties": { "type": { "const": "https://pgkronika.dev/problems/unknown-section" }, "status": { "const": 404 }, "params": { "$ref": "#/components/schemas/SectionParams" } } } },
          { "if": { "properties": { "code": { "const": "invalid_cursor" } } }, "then": { "properties": { "type": { "const": "https://pgkronika.dev/problems/invalid-cursor" }, "status": { "const": 400 }, "params": { "$ref": "#/components/schemas/EmptyParams" } } } },
          { "if": { "properties": { "code": { "const": "query_limit_exceeded" } } }, "then": { "properties": { "type": { "const": "https://pgkronika.dev/problems/query-limit-exceeded" }, "status": { "const": 413 }, "params": { "$ref": "#/components/schemas/LimitParams" } } } },
          { "if": { "properties": { "code": { "const": "analytic_capacity_unavailable" } } }, "then": { "properties": { "type": { "const": "https://pgkronika.dev/problems/analytic-capacity-unavailable" }, "status": { "const": 503 }, "params": { "$ref": "#/components/schemas/CapacityParams" } } } },
          { "if": { "properties": { "code": { "const": "store_read_failed" } } }, "then": { "properties": { "type": { "const": "https://pgkronika.dev/problems/store-read-failed" }, "status": { "const": 500 }, "params": { "$ref": "#/components/schemas/EmptyParams" } } } },
          { "if": { "properties": { "code": { "const": "internal_error" } } }, "then": { "properties": { "type": { "const": "https://pgkronika.dev/problems/internal-error" }, "status": { "const": 500 }, "params": { "$ref": "#/components/schemas/EmptyParams" } } } }
        ]
      },
...
    }
  }
}